Add hot_posts and developer_scores tables
From a schema perspective, the hot_posts view is just like the regular
posts table, with an additional column: score. Score is calculated using
a modified version of the Hacker News hot algorithm. The only difference
is that the gravity factor is decreased. This is because we only get a
handful of posts per day, rather than hundreds. Well-liked posts from a
few days ago tend to be a bit "sticky" with this lower gravity.

The developer_scores table combines the normal kinds of stats you might
want about each author (number of posts, total number of likes, average
number of likes per post), and combines it with the "hotness" factor.
Hotness is really just the sum of all of the scores from the hot_posts
view. However, the final value is scaled to a 1-to-10 linear scale.
The hottest developer will always have a score of 10; the nottest will
always have a score of 1.

Rather than simply linearly scale the raw sums of post scores, you'll
notice that we take the log2 of a scale of 2 to 1024, which, of course
results in a final 1-to-10 scale. This accounts for the fact that our
post score algorithm decays exponentially: without taking the logarithm,
the final "hotness" values would be all very close to 1, with the
exception of the highest. Using a logarithm here gives us a much more
even distribution of values. While this doesn't affect the sorting at
all, it does allow for the "hotness" value to be human-readable, and
potentially color coded, etc.